added 238, 239, and 240 element testing
