+++ pyansys/geometry.py
@@ -9,7 +9,7 @@ import numpy as np
 from pyansys import _relaxmidside, _reader
 from pyansys.misc import vtk_cell_info
 from pyansys.elements import ETYPE_MAP
-from pyansys._cellqual import cell_quality_float, cell_quality
+
 
 VTK9 = vtk.vtkVersion().GetVTKMajorVersion() >= 9
 INVALID_ALLOWABLE_TYPES = TypeError('`allowable_types` must be an array '
@@ -492,59 +492,6 @@ class Geometry():
         if self._node_angles is None:
             self._node_angles = np.ascontiguousarray(self._nodes[:, 3:])
         return self._node_angles
-
-    @property
-    def grid(self):
-        """VTK and pyvista grid of the archive file.
-
-        Examples
-        --------
-        >>> import pyansys
-        >>> archive = pyansys.Archive(pyansys.examples.hexarchivefile)
-        >>> archive.grid
-        UnstructuredGrid (0x7ffa237f08a0)
-          N Cells:      40
-          N Points:     321
-          X Bounds:     0.000e+00, 1.000e+00
-          Y Bounds:     0.000e+00, 1.000e+00
-          Z Bounds:     0.000e+00, 5.000e+00
-          N Arrays:     13
-        """
-        if self._grid is None:
-            raise AttributeError('Archive must be parsed as a vtk grid.\n'
-                                 'Set `parse_vtk=True`')
-        return self._grid
-
-    @property
-    def quality(self):
-        """Minimum scaled jacobian cell quality.
-
-        Negative values indicate invalid cells while positive values
-        indicate valid cells.  Varies between -1 and 1.
-
-        Examples
-        --------
-        >>> import pyansys
-        >>> archive = pyansys.Archive(pyansys.examples.hexarchivefile)
-        >>> archive.quality
-        array([1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
-               1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.,
-               1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.])
-        """
-        if self._grid is None:
-            raise AttributeError('Archive must be parsed as a vtk grid.\n'
-                                 'Set `parse_vtk=True`')
-        celltypes = self.grid.celltypes
-        points = self.grid.points
-        cells, offset = vtk_cell_info(self.grid)
-        if points.dtype == np.float64:
-            qual = cell_quality(cells, offset, celltypes, points)
-        else:
-            qual = cell_quality_float(cells, offset, celltypes, points)
-
-        # set qual of null cells to 1
-        qual[self._grid.celltypes == 0] = 1
-        return qual
 
     def __repr__(self):
         txt = 'ANSYS Geometry\n'
