+++ tests/elements/solid186/test_solid186.py
@@ -5,9 +5,7 @@ import numpy as np
 import pyansys
 from pyansys import examples
 
-
 test_path = os.path.dirname(os.path.abspath(__file__))
-testfiles_path = os.path.join(test_path, 'testfiles', 'solid186')
 
 
 @pytest.fixture(scope='module')
@@ -34,7 +32,7 @@ def test_geometry_nodenum(result, archive):
 
 
 def test_results_displacement(result):
-    textfile = os.path.join(testfiles_path, 'prnsol_u.txt')
+    textfile = os.path.join(test_path, 'prnsol_u.txt')
     nnum, r_values = result.nodal_solution(0)
     a_values = np.loadtxt(textfile, skiprows=2)[:, 1:4]
     assert np.allclose(r_values, a_values)
@@ -42,7 +40,7 @@ def test_results_displacement(result):
 
 def test_results_stress(result):
     _, r_values = result.nodal_stress(0)
-    textfile = os.path.join(testfiles_path, 'prnsol_s.txt')
+    textfile = os.path.join(test_path, 'prnsol_s.txt')
     a_values = np.loadtxt(textfile, skiprows=2)[:, 1:]
 
     # ignore nan
@@ -52,7 +50,7 @@ def test_results_stress(result):
 
 def test_results_pstress(result):
     r_nnum, r_values = result.principal_nodal_stress(0)
-    textfile = os.path.join(testfiles_path, 'prnsol_s_prin.txt')
+    textfile = os.path.join(test_path, 'prnsol_s_prin.txt')
     a_values = np.loadtxt(textfile, skiprows=2)[:, 1:]
 
     # ignore nan
